Can't kick if only two players

+++ back-end/db/player.ts
@@ -64,7 +64,13 @@ export default {
 
     kickPlayer: async (playerName: string, gameId: number): Promise<any> => {
       try {
-        await Player.deleteOne({name: playerName, gameId: gameId});
+        const players = await Player.find({gameId: gameId});
+        if (players.length > 2){
+          await Player.deleteOne({name: playerName, gameId: gameId});
+        }
+        else{
+          console.error('Need at least two players in game!');
+        }
       } catch (e) {
         console.error(`Issue kicking player: ${e}`);
       }
